Refactoring AdapterHelper
